rename responsive-grid files to auto-grid
