+++ public/app/features/dashboard-scene/scene/layouts-shared/paste.ts
@@ -9,14 +9,14 @@ import { LS_PANEL_COPY_KEY, LS_ROW_COPY_KEY, LS_TAB_COPY_KEY } from 'app/core/co
 import store from 'app/core/store';
 
 import { deserializeGridItem } from '../../serialization/layoutSerializers/DefaultGridLayoutSerializer';
-import { deserializeAutoGridItem } from '../../serialization/layoutSerializers/ResponsiveGridLayoutSerializer';
+import { deserializeAutoGridItem } from '../../serialization/layoutSerializers/AutoGridLayoutSerializer';
 import { deserializeRow } from '../../serialization/layoutSerializers/RowsLayoutSerializer';
 import { deserializeTab } from '../../serialization/layoutSerializers/TabsLayoutSerializer';
 import { dashboardSceneGraph } from '../../utils/dashboardSceneGraph';
 import { DashboardScene } from '../DashboardScene';
 import { DashboardGridItem } from '../layout-default/DashboardGridItem';
 import { GridCell } from '../layout-default/findSpaceForNewPanel';
-import { AutoGridItem } from '../layout-responsive-grid/ResponsiveGridItem';
+import { AutoGridItem } from '../layout-auto-grid/AutoGridItem';
 import { RowItem } from '../layout-rows/RowItem';
 import { TabItem } from '../layout-tabs/TabItem';
